Fix OpenACD settings to only allow update, remove extraneous functionality
